moved input binding
