+++ CoffeeEngine/src/CoffeeEngine/Scripting/Lua/Bindings/LuaInput.h
@@ -1,0 +1,6 @@
+#pragma once
+#include <sol/sol.hpp>
+
+namespace Coffee {
+    void RegisterInputBindings(sol::state& luaState);
+}
